+++ app/src/main/java/de/westnordost/streetmeasure/MeasureDisplayUnit.kt
@@ -5,7 +5,7 @@ import kotlin.math.round
 
 /** In which unit the measurement is displayed */
 sealed class MeasureDisplayUnit {
-    abstract fun format(distanceMeters: Float): String
+    abstract fun format(distanceMeters: Double): String
 }
 
 /** Measurement displayed in meters rounded to the nearest [cmStep] */
@@ -14,7 +14,7 @@ data class MeasureDisplayUnitMeter(val cmStep: Int) : MeasureDisplayUnit() {
         require(cmStep > 0)
     }
 
-    override fun format(distanceMeters: Float): String {
+    override fun format(distanceMeters: Double): String {
         val decimals = when {
             cmStep % 100 == 0 -> 0
             cmStep % 10 == 0 -> 1
@@ -24,7 +24,7 @@ data class MeasureDisplayUnitMeter(val cmStep: Int) : MeasureDisplayUnit() {
     }
 
     /** Returns the given distance in meters rounded to the given precision */
-    fun getRounded(distanceMeters: Float): Float {
+    fun getRounded(distanceMeters: Double): Double {
         return round(distanceMeters * 100 / cmStep) * cmStep / 100
     }
 }
@@ -35,13 +35,13 @@ data class MeasureDisplayUnitFeetInch(val inchStep: Int) : MeasureDisplayUnit() 
         require(inchStep in 1..12)
     }
 
-    override fun format(distanceMeters: Float): String {
+    override fun format(distanceMeters: Double): String {
         val (feet, inches) = getRounded(distanceMeters)
         return if (inches < 10) "$feet′ $inches″" else "$feet′$inches″"
     }
 
     /** Returns the given distance in meters as feet + inch */
-    fun getRounded(distanceMeters: Float): Pair<Int, Int> {
+    fun getRounded(distanceMeters: Double): Pair<Int, Int> {
         val distanceFeet = distanceMeters / 0.3048
         var feet = floor(distanceFeet).toInt()
         val inches = (distanceFeet - feet) * 12
